Table and Refresh Counter update
added design updates from Suede, created with cursor, mind the slop
