+++ src/components/tables/status.tsx
@@ -6,9 +6,6 @@ import { CircleCheck, CircleX, Loader2 } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
 
 export function StatusTable() {
-  const refreshInterval = 5 * 60;
-  const [countdown, setCountdown] = useState(refreshInterval);
-
   const { data: relayTestData, refetch, isLoading: isLoadingRelayTest, isFetching: isFetchingRelayTest } = useRelayTest();
   const { data: supplierData } = useSuppliers({
     paginationLimit: 1000,
@@ -25,51 +22,28 @@ export function StatusTable() {
     });
   }, [relayTestData, supplierData]);
 
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCountdown(countdown - 1);
-      if (countdown === 0) {
-        refetch();
-        setCountdown(refreshInterval);
-      }
-    }, 1000);
-    return () => clearInterval(interval);
-  }, [countdown, refetch]);
-
   return (
     <div className="flex flex-col gap-2">
-      <div className="flex justify-end items-end mb-2 gap-2 mr-2">
-        {isLoadingRelayTest || isFetchingRelayTest ? (
-          <Loader2 className="animate-spin w-4 h-4" />
-        ) : (
-          <div className="flex flex-col gap-2 text-sm">
-            <div className="flex justify-start gap-4">
-              <p className="font-semibold">Refreshed in:</p>
-              <p className="min-w-10 text-right">{countdown}s</p>
-            </div>
-          </div>
-        )}
-      </div>
       <Table>
         <TableHeader>
           <TableRow>
             <TableHead>Chain</TableHead>
-            <TableHead># of Suppliers</TableHead>
-            <TableHead>Status</TableHead>
-            <TableHead>Latency</TableHead>
-            <TableHead>Block</TableHead>
+            <TableHead className="text-center">Supplier Count</TableHead>
+            <TableHead className="text-center">Status</TableHead>
+            <TableHead className="text-center">Latency</TableHead>
+            <TableHead className="text-center">Block</TableHead>
           </TableRow>
         </TableHeader>
         <TableBody>
           {data && data.length > 0 ? data.map((item) => (
             <TableRow key={item.chain}>
               <TableCell>{item.chain}</TableCell>
-              <TableCell>{item.supplier ?? "--"}</TableCell>
-              <TableCell>
+              <TableCell className="text-center">{item.supplier ?? "--"}</TableCell>
+              <TableCell className="text-center">
                 <HealthLabel status={item.status} />
               </TableCell>
-              <TableCell>{item.latency ?? "--"}ms</TableCell>
-              <TableCell>{item.blockNumber ?? "--"}</TableCell>
+              <TableCell className="text-center">{item.latency ?? "--"}ms</TableCell>
+              <TableCell className="text-center">{item.blockNumber ?? "--"}</TableCell>
             </TableRow>
           )) : (
             <TableRow>
@@ -84,7 +58,7 @@ export function StatusTable() {
 
 export function HealthLabel({ status }: { status: string }) {
   return (
-    <div className={cn(status === "success" ? "text-[#60BC29]" : status === "error" ? "text-[#E03834]" : "text-yellow-500")}>
+    <div className={cn("flex justify-center", status === "success" ? "text-[#60BC29]" : status === "error" ? "text-[#E03834]" : "text-yellow-500")}>
       {status === "success" ? (
         <div className="flex items-center gap-2">
           <CircleCheck className="w-4 h-4" />
